Upgrade to HTTP/2 draft 13
Motivation:

Need to upgrade HTTP/2 implementation to latest draft.

Modifications:

Various changes to support draft 13.

Result:

Support for HTTP/2 draft 13.

+++ codec-http2/src/main/java/io/netty/handler/codec/http2/Http2FrameReader.java
@@ -36,12 +36,12 @@ public interface Http2FrameReader extends Closeable {
     /**
      * Sets the maximum size of the HPACK header table used for decoding HTTP/2 headers.
      */
-    void maxHeaderTableSize(int max);
+    void maxHeaderTableSize(long max);
 
     /**
      * Gets the maximum size of the HPACK header table used for decoding HTTP/2 headers.
      */
-    int maxHeaderTableSize();
+    long maxHeaderTableSize();
 
     /**
      * Closes this reader and frees any allocated resources.
